Changelog: Added LinearSystem to replace LinearEventSystem, accepts async and blocking calls, where the blocking calls require a Terminator
Added Path and PathBuilder to create and build dynamic paths. WARNING: Line implementation is stable, but the stability of splines cannot be guarenteed

Added CrosstrackPathFollower to follow paths using a guided crosstrack following algorithm.

Added more RobotConstants for velocity constraints, Angle.java to prevent angle confusion, and ProgramClock for a universal program timer.

Added a default curve approximation method and a method to calculate path length in MathUtils, as well as the ability to subtract and rotate Vector2

Added RedAutonomous3, which is a copy of RedAutonomous2 with the 1 ring path switched to the new system

Moved packages around to clean up files